+++ src/index.js
@@ -1,6 +1,6 @@
 import { BreakpointContainer, BrowserContainer } from './components/BreakpointContainer';
 import { withBreakpointContainer, withBrowserContainer } from './components/HOCs';
-import { Breakpoint } from './components/Breakpoint';
+import Breakpoint from './components/Breakpoint';
 import { BREAKPOINTS, getBpUpperLimit } from './data/breakpoints';
 
 export {